requirements.txt
mysql
python-dotenv
mysql-connector-python
